Added some extra comments for clarity
start(): added variable continue_executing to allow user to return to file_manipulator if desired, changed the while loop from True to continue_executing, changed if statements from looking into a tuple to comparing the .casefold() to ‘x’.
directory_crawler(): Removed call to exit_program(), this forces the return to start() for program further execution, and a call to exit_program giving user a chance to return to the main menu.
exit_program(): Removed unneeded else, Change the return from None to False.